+++ lib/karafka/setup/config.rb
@@ -4,10 +4,19 @@ module Karafka
   module Setup
     # Configurator for setting up all the framework details that are required to make it work
     # @note If you want to do some configurations after all of this is done, please add to
-    #   karafka/config a proper file (needs to inherit from Karafka::Setup::Config::Base
+    #   karafka/config a proper file (needs to inherit from Karafka::Setup::Configurators::Base
     #   and implement setup method) after that everything will happen automatically
-    # @see Karafka::Setup::Config::Base for more details about configurators api
+    # @note This config object allows to create a 1 level nestings (nodes) only. This should be
+    #   enough and will still keep the code simple
+    # @see Karafka::Setup::Configurators::Base for more details about configurators api
     class Config
+      # Nested settings node (for hash based settings) - this will allow us to have nice
+      # node fetching like kafka.hosts instead of kafka[:hosts]
+      # @example Simple node for kafka hosts
+      #   node = Node.new(['127.0.0.1'])
+      #   node.hosts #=> ['127.0.0.1']
+      class Node < OpenStruct; end
+
       class << self
         attr_accessor :config
       end
@@ -19,8 +28,8 @@ module Karafka
       # option name [String] current app name - used to provide default Kafka groups namespaces
       # option redis [Hash] redis options hash (url and optional parameters)
       # option wait_timeout [Integer] seconds that we will wait on a single topic for messages
-      # option zookeeper_hosts [Array] zookeeper hosts with ports where zookeeper servers are run
-      # option kafka_hosts [Array] - optional - kafka hosts with ports (autodiscovered if missing)
+      # option zookeeper [Hash] zookeeper configuration options (hosts with ports and chroot)
+      # option kafka [Hash] - optional - kafka configuration options (hosts)
       SETTINGS = [
         :logger,
         :max_concurrency,
@@ -28,11 +37,15 @@ module Karafka
         :name,
         :redis,
         :wait_timeout,
-        :zookeeper_hosts,
-        :kafka_hosts
+        zookeeper: %i( hosts ),
+        kafka: %i( hosts )
       ].freeze
 
-      SETTINGS.each do |attr_name|
+      # We assume that we can have up to two levels of settings (root settings + additional node)
+      # No more nodes are supported (1 level nestings)
+      ROOT_SETTINGS = SETTINGS.flat_map { |set| set.is_a?(Hash) ? set.keys : set }
+
+      ROOT_SETTINGS.each do |attr_name|
         attr_writer attr_name
 
         # @return A given setting value if defined or a default one if not defined
@@ -43,11 +56,12 @@ module Karafka
         #   logger #=> Karafka::Logger instance
         define_method attr_name do
           value = instance_variable_get(:"@#{attr_name}")
+          value ||= Defaults.public_send(attr_name) if Defaults.respond_to?(attr_name)
 
-          return value if value
-          return Defaults.public_send(attr_name) if Defaults.respond_to?(attr_name)
-
-          nil
+          # If a value is a simple type - we will just return it, if it was defined as a hash
+          # (for example for kafka and zookeeper) - we will return a Node object that will allow
+          # us to have a nicer fetching
+          value.is_a?(Hash) ? Node.new(value) : value
         end
       end